🚚 [Move] 디렉토리 위치 수정
utils, constants, services 디렉토리 위치 최상위로 수정
