+++ services/versions.js
@@ -1,10 +1,9 @@
-import generateApiUri from "../components/utils/generateURI";
+import generateApiUri from "../utils/generateURI";
 
 const baseURI = import.meta.env.VITE_BACKEND_BASE_API_URI;
 
 const getVersions = async projectsId => {
   const token = JSON.parse(localStorage.getItem("FigmaToken")).access_token;
-
   const API_URI = generateApiUri(baseURI, `projects/${projectsId}/versions`);
 
   const response = await fetch(API_URI, {
@@ -15,7 +14,7 @@ const getVersions = async projectsId => {
   });
 
   const responseJson = await response.json();
-  const { versions } = responseJson;
+  const versions = responseJson.content;
 
   return versions;
 };
