+++ utils/getProjectKeyFromURI.js
